requirements.txt
paho-mqtt==1.6.1
pytest==7.4.2
python-dotenv==1.0.0